Fix numbering of words (#13)
* Fix numbering of words

The algorithm used to look for black squares to the left or above a
possible word.  This patch checks that the resulting word is longer
than one character.

Signed-off-by: Peter Chubb <peter@chubb.wattle.id.au>

* Simplify logic, fix _across and _down arrays

The previous change fixed the numbers, but didn't adjust
the _across and _down arrays to be correct.

SIgned-off-by: Peter Chubb <peter@chubb.wattle.id.au>

* Fix fill() with unchecked lights
Make Crossword.get_word_at_index() return None if there is
no word.  Make the matching code cope with None returns.

Signed-off-by: Peter Chubb <peter@chubb.wattle.id.au>

Signed-off-by: Peter Chubb <peter@chubb.wattle.id.au>